backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.12.0
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5
swagger-ui-bundle==1.1.0
